+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/reporter/SpotbugsReporterTest.groovy
@@ -40,11 +40,11 @@ class SpotbugsReporterTest extends AbstractKitTest {
         error.replaceAll("\r", '').contains """
 2 (0 / 2 / 0) SpotBugs violations were found in 2 files
 
-[Performance | URF_UNREAD_FIELD] sample.(Sample.java:11)  [priority 2]
+[Performance | URF_UNREAD_FIELD] sample.(Sample.java:11)  [priority 2 / rank 18]
 \t>> Unread field: sample.Sample.sample
   This field is never read. Consider removing it from the class.
 
-[Performance | URF_UNREAD_FIELD] sample.(Sample2.java:11)  [priority 2]
+[Performance | URF_UNREAD_FIELD] sample.(Sample2.java:11)  [priority 2 / rank 18]
 \t>> Unread field: sample.Sample2.sample
   This field is never read. Consider removing it from the class.