+++ src/test/resources/ru/vyarus/gradle/plugin/quality/java/sample/Sample4.java
@@ -1,0 +1,22 @@
+package sample;
+
+import java.util.ArrayDeque;
+import java.util.Deque;
+
+public class Sample4 {
+
+    private String sample;
+
+    public Sample4(String sample) {
+        this.sample = sample;
+    }
+
+    public static void main(String[] args) {
+        final Deque res = new ArrayDeque();
+        System.out.println(res);
+    }
+
+    public Boolean someting() {
+        return null;
+    }
+}